fully function TabuSearchSVRP, but does not found feasible solutions when number of vehicles is low

+++ include/TabuSearchSVRP.h
@@ -91,7 +91,7 @@ struct svrpSol {
 class TabuSearchSVRP {
 
     Graph g;
-    double penalty, bestPenalExpCost;
+    long double penalty, bestPenalExpCost;
     int numVehicles, capacity, numSelected, numNearest, numRoutes;
     int itCount, numInfeasibleNearby;
     int currNoImprovement, maxNoImprovement;
